refact(header): extract DiscordButton to its own component
